refactor: move run() out of bin.ts and drop file-wide lint disables
Addresses review feedback: bin.ts stays under the max-lines and
max-dependencies limits without blanket suppressions.
