+++ src/bin.ts
@@ -1,5 +1,3 @@
-/* eslint-disable max-lines */
-/* oxlint-disable import/max-dependencies */
 import fs from 'node:fs';
 import colors from 'yoctocolors';
 import enquirer from 'enquirer';
@@ -7,7 +5,6 @@ import fuzzysearch from 'fuzzysearch';
 import mri from 'mri';
 import glob from 'tiny-glob/sync.js';
 import { parse } from './domain/repo.js';
-import degit from './index.js';
 import {
 	handleAliasSubcommand,
 	handleListSubcommand,
@@ -15,7 +12,7 @@ import {
 	loadAliases,
 	resolveAlias,
 } from './aliases.js';
-import { startSpinner } from './shared/spinner.js';
+import { run } from './cli/run.js';
 import { base, DegitError, tryReadJson } from './shared/utils.js';
 
 type Choice = {
@@ -44,16 +41,6 @@ type PromptResult = {
 
 type ForceResult = {
 	force: boolean;
-};
-
-type RunArgs = {
-	aliases?: Record<string, string>;
-	cache?: boolean;
-	files?: string[];
-	force?: boolean;
-	interactive?: boolean;
-	mode?: string;
-	verbose?: boolean;
 };
 
 /* eslint-disable security/detect-non-literal-fs-filename */
@@ -242,67 +229,7 @@ export async function main(argv: string[]) {
 }
 
 /* eslint-enable security/detect-non-literal-fs-filename */
-export function run(src: string, dest: string, args: RunArgs) {
-	const { interactive, ...degitArgs } = args;
-	const d = degit(src, degitArgs as Parameters<typeof degit>[1]);
-	const spinner = interactive ? startSpinner() : null;
-
-	d.on('info', (event) => {
-		if (event.code === 'SUCCESS') {
-			spinner?.stop();
-		} else {
-			spinner?.clear();
-		}
-		console.log(colors.cyan(`> ${event.message.replace('options.', '--')}`));
-	});
-
-	d.on('warn', (event) => {
-		spinner?.clear();
-		console.warn(colors.magenta(`! ${event.message.replace('options.', '--')}`));
-	});
-
-	d.clone(dest)
-		.then(() => spinner?.stop())
-		.catch((error: Error) => {
-			spinner?.stop();
-			console.error(colors.red(`! ${error.message.replace('options.', '--')}`));
-			if (args.verbose) {
-				const detail = getCloneErrorDetail(error);
-
-				if (detail) {
-					console.error(detail);
-				}
-			}
-			process.exit(1);
-		});
-}
-
-function getCloneErrorDetail(error: unknown): string | undefined {
-	if (!error || typeof error !== 'object') {
-		return undefined;
-	}
-
-	const nestedError =
-		'original' in error ? error.original : 'cause' in error ? error.cause : undefined;
-
-	if (!nestedError) {
-		return undefined;
-	}
-
-	if (nestedError instanceof Error) {
-		return nestedError.stack || nestedError.message;
-	}
-
-	if (typeof nestedError === 'string') {
-		return nestedError;
-	}
-
-	try {
-		return JSON.stringify(nestedError);
-	} catch {
-		return String(nestedError);
-	}
-}
+export { run };
 
 if (!process.env.VITEST) {
 	try {
